refactor: move buildkit connection code to new package
This is the first part of refactoring this logic, preparing it for
migration to the new CLI, structuring a clean interface for future
extension, and preparing to replace the testcontainers approach.
