+++ buildkit/buildkit.go
@@ -1,0 +1,3 @@
+// The buildkit package provides various utilities and functions for working
+// with BuildKit.
+package buildkit
