+++ initrd/dockerfile.go
@@ -9,16 +9,15 @@ import (
 	"encoding/csv"
 	"fmt"
 	"io"
-	"net"
 	"os"
 	"path/filepath"
-	"runtime/debug"
 	"strings"
 
 	"golang.org/x/sync/errgroup"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
+	"kraftkit.sh/buildkit"
 	"kraftkit.sh/cmdfactory"
 	"kraftkit.sh/config"
 	"kraftkit.sh/fs/cpio"
@@ -35,9 +34,6 @@ import (
 	"github.com/moby/buildkit/session/secrets/secretsprovider"
 	"github.com/moby/buildkit/session/sshforward/sshprovider"
 	"github.com/moby/buildkit/util/progress/progressui"
-	"github.com/testcontainers/testcontainers-go"
-	tlog "github.com/testcontainers/testcontainers-go/log"
-	"github.com/testcontainers/testcontainers-go/wait"
 
 	_ "github.com/moby/buildkit/client/connhelper/dockercontainer"
 	_ "github.com/moby/buildkit/client/connhelper/kubepod"
@@ -89,73 +85,6 @@ func init() {
 				"Supply secrets when building Dockerfile",
 			),
 		)
-	}
-}
-
-var testcontainersLoggingHook = func(logger tlog.Logger) testcontainers.ContainerLifecycleHooks {
-	shortContainerID := func(c testcontainers.Container) string {
-		return c.GetContainerID()[:12]
-	}
-
-	return testcontainers.ContainerLifecycleHooks{
-		PreCreates: []testcontainers.ContainerRequestHook{
-			func(ctx context.Context, req testcontainers.ContainerRequest) error {
-				logger.Printf("creating container for image %s", req.Image)
-				return nil
-			},
-		},
-		PostCreates: []testcontainers.ContainerHook{
-			func(ctx context.Context, c testcontainers.Container) error {
-				logger.Printf("container created: %s", shortContainerID(c))
-				return nil
-			},
-		},
-		PreStarts: []testcontainers.ContainerHook{
-			func(ctx context.Context, c testcontainers.Container) error {
-				logger.Printf("starting container: %s", shortContainerID(c))
-				return nil
-			},
-		},
-		PostStarts: []testcontainers.ContainerHook{
-			func(ctx context.Context, c testcontainers.Container) error {
-				logger.Printf("container started: %s", shortContainerID(c))
-				return nil
-			},
-		},
-		PreStops: []testcontainers.ContainerHook{
-			func(ctx context.Context, c testcontainers.Container) error {
-				logger.Printf("stopping container: %s", shortContainerID(c))
-				return nil
-			},
-		},
-		PostStops: []testcontainers.ContainerHook{
-			func(ctx context.Context, c testcontainers.Container) error {
-				logger.Printf("container stopped: %s", shortContainerID(c))
-				return nil
-			},
-		},
-		PreTerminates: []testcontainers.ContainerHook{
-			func(ctx context.Context, c testcontainers.Container) error {
-				logger.Printf("terminating container: %s", shortContainerID(c))
-				return nil
-			},
-		},
-		PostTerminates: []testcontainers.ContainerHook{
-			func(ctx context.Context, c testcontainers.Container) error {
-				logger.Printf("container terminated: %s", shortContainerID(c))
-				return nil
-			},
-		},
-	}
-}
-
-type testcontainersPrintf struct {
-	ctx context.Context
-}
-
-func (t *testcontainersPrintf) Printf(format string, v ...interface{}) {
-	if config.G[config.KraftKit](t.ctx).Log.Level == "trace" {
-		log.G(t.ctx).Tracef(format, v...)
 	}
 }
 
@@ -216,43 +145,6 @@ func (initrd *dockerfile) Name() string {
 	return "Dockerfile"
 }
 
-func startBuildkit(ctx context.Context, buildkitVersion string, port int, printf *testcontainersPrintf) (testcontainers.Container, error) {
-	// Trap any panics that occur when instantiating BuildKit through the
-	// testcontainers library. This is known happen if Docker is not installed.
-	// For more information see:
-	//
-	// https://github.com/unikraft/kraftkit/issues/2001
-	defer func() {
-		if r := recover(); r != nil {
-			log.G(ctx).Warn("could not start BuildKit ephemeral container")
-			log.G(ctx).Warn("this can be caused by Docker is either not running or inaccessible")
-			log.G(ctx).Warn("")
-			log.G(ctx).Warn("if you think this was caused by something else, please open an issue at:")
-			log.G(ctx).Warn("https://github.com/unikraft/kraftkit/issues")
-		}
-	}()
-	return testcontainers.GenericContainer(ctx, testcontainers.GenericContainerRequest{
-		Started: true,
-		Logger:  printf,
-		ContainerRequest: testcontainers.ContainerRequest{
-			AlwaysPullImage: true,
-			Image:           "moby/buildkit:" + buildkitVersion,
-			WaitingFor:      wait.ForLog(fmt.Sprintf("running server on [::]:%d", port)),
-			Privileged:      true,
-			ExposedPorts:    []string{fmt.Sprintf("%d:%d/tcp", port, port)},
-			Cmd:             []string{"--addr", fmt.Sprintf("tcp://0.0.0.0:%d", port)},
-			Mounts: testcontainers.ContainerMounts{
-				{
-					Source: testcontainers.GenericVolumeMountSource{
-						Name: "kraftkit-buildkit-cache",
-					},
-					Target: "/var/lib/buildkit",
-				},
-			},
-		},
-	})
-}
-
 // Build implements Initrd.
 func (initrd *dockerfile) Build(ctx context.Context) (string, error) {
 	if initrd.opts.output == "" {
@@ -278,95 +170,13 @@ func (initrd *dockerfile) Build(ctx context.Context) (string, error) {
 	defer ociOutput.Close()
 	defer os.RemoveAll(ociOutput.Name())
 
-	buildkitAddr := config.G[config.KraftKit](ctx).BuildKitHost
-	c, _ := client.New(ctx, buildkitAddr)
-	buildKitInfo, connerr := c.Info(ctx)
-	if connerr != nil {
-		log.G(ctx).Info("creating ephemeral buildkit container")
-
-		buildkitVersion := "latest"
-		if bi, ok := debug.ReadBuildInfo(); ok {
-			for _, dep := range bi.Deps {
-				if dep.Path == "github.com/moby/buildkit" {
-					buildkitVersion = dep.Version
-					break
-				}
-			}
-			log.G(ctx).Debug("could not determine BuildKit version from module list")
-		}
-
-		testcontainers.DefaultLoggingHook = testcontainersLoggingHook
-		printf := &testcontainersPrintf{ctx}
-		tlog.SetDefault(printf)
-
-		// Trap any errors with a helpful message for how to use buildkit
-		defer func() {
-			if connerr == nil {
-				return
-			}
-
-			log.G(ctx).Warnf("could not connect to BuildKit client '%s' is BuildKit running?", buildkitAddr)
-			log.G(ctx).Warn("")
-			log.G(ctx).Warn("By default, KraftKit will look for a native install which")
-			log.G(ctx).Warn("is located at /run/buildkit/buildkit.sock.  Alternatively, you")
-			log.G(ctx).Warn("can run BuildKit in a container (recommended for macOS users)")
-			log.G(ctx).Warn("which you can do by running:")
-			log.G(ctx).Warn("")
-			log.G(ctx).Warn("  docker run --rm -d --name buildkit --privileged moby/buildkit:" + buildkitVersion)
-			log.G(ctx).Warn("")
-			log.G(ctx).Warn("Depending on your container runtime, you should connect to Buildkit via:")
-			log.G(ctx).Warn("")
-			log.G(ctx).Warn("  export KRAFTKIT_BUILDKIT_HOST=docker-container://buildkit # for docker")
-			log.G(ctx).Warn("or")
-			log.G(ctx).Warn("  export KRAFTKIT_BUILDKIT_HOST=podman-container://buildkit # for podman")
-			log.G(ctx).Warn("")
-			log.G(ctx).Warn("For more usage instructions visit: https://unikraft.org/buildkit")
-			log.G(ctx).Warn("")
-		}()
-
-		// Port 0 means "give me any free port"
-		addr, err := net.ResolveTCPAddr("tcp", ":0")
-		if err != nil {
-			return "", err
-		}
-		l, err := net.ListenTCP("tcp", addr)
-		if err != nil {
-			return "", err
-		}
-
-		port := l.Addr().(*net.TCPAddr).Port
-		_ = l.Close()
-
-		buildkitd, err := startBuildkit(ctx, buildkitVersion, port, printf)
-		if err != nil {
-			return "", fmt.Errorf("creating buildkit container: %w", err)
-		}
-
-		if buildkitd == nil {
-			return "", fmt.Errorf("could not start ephemeral BuildKit container")
-		}
-
-		defer func() {
-			if err := buildkitd.Terminate(ctx); err != nil && !strings.Contains(err.Error(), "context cancelled") {
-				log.G(ctx).
-					WithError(err).
-					Debug("terminating buildkit container")
-			}
-		}()
-
-		buildkitAddr = fmt.Sprintf("tcp://localhost:%d", port)
-
-		c, _ = client.New(ctx, buildkitAddr)
-		buildKitInfo, connerr = c.Info(ctx)
-		if err != nil {
-			return "", fmt.Errorf("connecting to container buildkit client: %w", err)
-		}
+	c, cleanup, err := buildkit.ConnectToBuildkit(ctx)
+	if err != nil {
+		return "", fmt.Errorf("could not connect to buildkit: %w", err)
 	}
-
-	log.G(ctx).
-		WithField("addr", buildkitAddr).
-		WithField("version", buildKitInfo.BuildkitVersion.Version).
-		Debug("using buildkit")
+	if cleanup != nil {
+		defer cleanup()
+	}
 
 	var cacheExports []client.CacheOptionsEntry
 	if len(initrd.opts.cacheDir) > 0 {
